Add initial implementation of AsyncModel
